+++ desc/objectives/_fast_ion.py
@@ -171,6 +171,7 @@ class GammaC(_Objective):
         surf_batch_size=1,
         spline=False,
         Nemov=True,
+        device_id=0,
     ):
         if target is None and bounds is None:
             target = 0.0
@@ -207,6 +208,7 @@ class GammaC(_Objective):
             deriv_mode=deriv_mode,
             name=name,
             jac_chunk_size=jac_chunk_size,
+            device_id=device_id,
         )
 
     def build(self, use_jit=True, verbose=1):